added CreateFoamTypeObjects function to foamunitpriceservice
now when you create a FoamUnitPriceService object, you pass through a csvfilepath, it is read and turned into a list of string arrays , a new list of foam objects is created and each object's fields are populated with the values of each row. One row in the csv for one object. FoamUnitPriceServce._listOfFoamTypeObjects can be called to access the list of objects anywhere in the code.